+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 This log should always be updated when doing backwards incompatible changes, resulting in a major version bump. Feel free to add a log for lesser version bumps as well, but for major bumps it's a must.
 
+**unreleased major**
+-----
+- Changed
+    - `transactions.Transaction` init only takes keyword arguments.
+
 2.0.0
 -----
 - Changed